# m8flow-backend/src/m8flow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: m8flow-backend-extension
paths:
...
  /templates:
    get:
      summary: List templates
      description: Retrieves a list of templates with optional filtering by latest version, category, tag, owner, visibility, and search query
      operationId: m8flow_backend.routes.templates_controller.template_list
      tags:
        - Template
      parameters:
        - name: latest_only
          in: query
          schema:
            type: boolean
          description: Filter to show only the latest version of each template
        - name: category
          in: query
          schema:
            type: string
          description: Filter templates by category
        - name: tag
          in: query
          schema:
            type: string
          description: Filter templates by tag
        - name: owner
          in: query
          schema:
            type: string
          description: Filter templates by owner
        - name: visibility
          in: query
          schema:
            type: string
            enum: [PRIVATE, TENANT, PUBLIC]
          description: Filter templates by visibility level
        - name: search
          in: query
          schema:
            type: string
          description: Search templates by name or description
        - name: template_key
          in: query
          schema:
            type: string
          description: Filter to this template key (returns all versions if latest_only=false)
        - name: published_only
          in: query
          schema:
            type: boolean
          description: When true, return only published templates
        - name: include_deleted
          in: query
          schema:
            type: boolean
          description: When true, include soft-deleted templates in results
        - name: deleted_only
          in: query
          schema:
            type: boolean
          description: When true, return only soft-deleted templates
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created, name]
          description: Sort results by created date or name
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (asc or desc)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (1-indexed)
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
          description: Items per page (max 100)
      responses:
        "200":
          description: List of templates retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResponse"
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Create template
      description: Creates a new template using BPMN XML content with metadata provided via X-Template-* headers.
      operationId: m8flow_backend.routes.templates_controller.template_create
      tags:
        - Template
      parameters:
        - name: X-Template-Key
          in: header
          required: true
          schema:
            type: string
          description: Unique key for the template.
        - name: X-Template-Name
          in: header
          required: true
          schema:
            type: string
          description: Human-readable name of the template.
        - name: X-Template-Description
          in: header
          required: false
          schema:
            type: string
          description: Optional description of the template.
        - name: X-Template-Category
          in: header
          required: false
          schema:
            type: string
          description: Optional category for grouping templates.
        - name: X-Template-Tags
          in: header
          required: false
          schema:
            type: string
          description: JSON array string or comma-separated list of tags.
        - name: X-Template-Visibility
          in: header
          required: false
          schema:
            type: string
            enum: [PRIVATE, TENANT, PUBLIC]
          description: Visibility level of the template. Defaults to PRIVATE.
        - name: X-Template-Status
          in: header
          required: false
          schema:
            type: string
          description: Status of the template. Defaults to draft.
        - name: X-Template-Is-Published
          in: header
          required: false
          schema:
            type: string
            enum: ["true", "false"]
          description: Whether the template is published. Defaults to false.
        - name: X-Template-Version
          in: header
          required: false
          schema:
            type: string
          description: Optional explicit version identifier (e.g., V1). If omitted, the next version will be generated automatically.
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
              description: BPMN XML content (single file)
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              description: Multiple files (.bpmn, .json, .dmn, .md). At least one .bpmn required. Metadata via X-Template-* headers.
      responses:
        "201":
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "400":
          description: Bad request - missing required fields or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - template with the same key and version already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /templates/{id}:
    get:
      summary: Get template by ID
      description: Retrieves a specific template by its unique identifier
      operationId: m8flow_backend.routes.templates_controller.template_get_by_id
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Unique identifier of the template
        - name: include_contents
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Whether to include BPMN content in the response. Set to false for better performance when only metadata is needed.
        - name: include_deleted
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether to include soft-deleted templates. Set to true to fetch a deleted template's details.
      responses:
        "200":
          description: Template retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update template
      description: Updates an existing template's metadata and optionally its BPMN content. Cannot update template_key or version
      operationId: m8flow_backend.routes.templates_controller.template_update_by_id
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Unique identifier of the template to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdateRequest"
          application/xml:
            schema:
              type: string
              description: Updated BPMN XML content
      responses:
        "200":
          description: Template updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "400":
          description: Bad request - no fields to update or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete template
      description: >
        State-aware delete behavior:
        draft templates are hard-deleted (creator or tenant-admin);
        published templates are soft-deleted with a renamed name suffix
        (tenant-admin only).
      operationId: m8flow_backend.routes.templates_controller.template_delete_by_id
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Unique identifier of the template to delete
      responses:
        "200":
          description: Template deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request - invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - user not allowed to delete this template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /templates/{id}/restore:
    post:
      summary: Restore deleted template
      description: Restores a soft-deleted template (tenant-admin only)
      operationId: m8flow_backend.routes.templates_controller.template_restore_by_id
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Unique identifier of the template to restore
      responses:
        "200":
          description: Template restored successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "400":
          description: Bad request - template is not deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - tenant-admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /templates/{id}/create-process-model:
    post:
      summary: Create a process model from a template
      operationId: m8flow_backend.routes.templates_controller.template_create_process_model
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Template ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessModelFromTemplateRequest"
      responses:
        "201":
          description: Process model created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProcessModelFromTemplateResponse"
        "400":
          description: Bad request - missing required fields or template version is not published
        "401":
          description: Unauthorized
        "404":
          description: Template not found
...
  /templates/{id}/bpmn:
    get:
      summary: Get BPMN XML
      description: Retrieves the BPMN XML content associated with a specific template
      operationId: m8flow_backend.routes.templates_controller.template_get_bpmn
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Unique identifier of the template
      responses:
        "200":
          description: BPMN XML retrieved successfully
          content:
            application/xml:
              schema:
                type: string
                description: BPMN 2.0 XML content
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Template or BPMN content not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: integer
        templateKey:
          type: string
        version:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        tenantId:
          type: string
        visibility:
          type: string
          enum: [PRIVATE, TENANT, PUBLIC]
        files:
          type: array
          items:
            type: object
            properties:
              fileType:
                type: string
                enum: [bpmn, json, dmn, md]
              fileName:
                type: string
        isPublished:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        modifiedBy:
          type: string
        bpmnContent:
          type: string
        is_deleted:
          type: boolean
    TemplateListResponse:
      type: array
      items:
        $ref: "#/components/schemas/Template"
...
    TemplateUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        visibility:
          type: string
          enum: [PRIVATE, TENANT, PUBLIC]
        status:
          type: string
...
    MessageResponse:
      $ref: "#/components/schemas/MessageData"
...
    ErrorResponse:
      $ref: "#/components/schemas/ErrorData"
...
    CreateProcessModelFromTemplateResponse:
      type: object
      properties:
        process_model:
          type: object
          description: The created process model info
        template_info:
          $ref: "#/components/schemas/ProcessModelTemplateInfo"
...